Replace Functions with built-in methods
Functions are underused, and no longer editable in QubyAdmin.  The
helper-methods themselves are used quite often, but they never change,
and there are only three (really only two, even).

We can build these functions in as helper methods that are available in
the questionnaire builder dsl, removing the need for functions to exist
as files on disk.