Replace Functions with built-in methods
Functions are underused, and no longer editable in QubyAdmin.  The
helper-methods themselves are used quite often, but they never change,
and there are only three (really only two, even).

We can build these functions in as helper methods that are available in
the questionnaire builder dsl, removing the need for functions to exist
as files on disk.

Closes #65

LGTM given by: @jornvandebeek

Squashed commit of the following:

commit fbc7462b4b41b174967784138af7f42dddc7ed85
Author: Marten Veldthuis <marten@veldthuis.com>
Date:   Wed Apr 16 20:14:04 2014 +0200

    Disable line length rubocop for this file

commit ab5e4d483db0cbb7fd0009c0c893731108340342
Author: Marten Veldthuis <marten@veldthuis.com>
Date:   Wed Apr 16 18:43:38 2014 +0200

    Replace Functions with built-in methods

    Functions are underused, and no longer editable in QubyAdmin.  The
    helper-methods themselves are used quite often, but they never change,
    and there are only three (really only two, even).

    We can build these functions in as helper methods that are available in
    the questionnaire builder dsl, removing the need for functions to exist
    as files on disk.